refactor: 코드 패키지 리팩토링
presentation -> API
application -> Service
